requirements.txt
# Multi-Level Flight Control - Core Dependencies
# Install with: pip install -r requirements.txt

# Core Scientific Computing
numpy>=1.20.0,<2.0.0
scipy>=1.7.0
dataclasses-json>=0.5.0
pyyaml>=6.0.0

# Geodetic Conversion (for accurate NED coordinates)
pyproj>=3.3.0

# Visualization
matplotlib>=3.4.0
plotly>=5.0.0
pyvista>=0.35.0
kaleido>=0.2.0  # For static image export from plotly

# Interactive GUI
pygame>=2.1.0

# Data Handling & Logging
pandas>=1.3.0
h5py>=3.0.0

# Reinforcement Learning (optional - for learned controllers)
stable-baselines3[extra]>=2.0.0
sb3-contrib>=2.0.0  # For RecurrentPPO
gymnasium>=0.28.0
torch>=1.13.0  # PyTorch (CPU or GPU version)

# Experiment Tracking
tensorboard>=2.11.0

# C++ Bindings
pybind11>=2.9.0

# Physics Validation (optional - for JSBSim comparison)
jsbsim>=1.1.0

# Hardware Interface (optional - for Phase 6)
pyserial>=3.5
